Split abi.h into SYS_* dispatch numbers and a new linux-wire.h
syscall/abi.h carried two unrelated things: the SYS_* dispatch numbers,
and the Linux wire ABI (errno values, flag constants, wire-format
structs, the FD table).

Almost every syscall implementation file needs the wire half and never
touches SYS_* at all, so a change to a syscall number forced a rebuild of
files that only wanted a struct layout, and vice versa. The wire half
moves to syscall/linux-wire.h and abi.h keeps the numbers plus the two
dispatch entry points.

The SYS_* numbers have exactly two readers after the split: syscall.c,
and debug/syscall-hist.c through the generated dispatch.h. The other
files that still include abi.h (core/bootstrap.c, runtime/fork-state.c,
syscall/proc.c) want only the entry-point declarations, and include
linux-wire.h separately for the wire half.

Mechanical apart from the move: every other change here is an include
line.
