+++ src/syscall/mem.c
@@ -24,7 +24,7 @@
 #include "utils.h"
 
 #include "runtime/thread.h"
-#include "syscall/abi.h"
+#include "syscall/linux-wire.h"
 #include "syscall/fuse.h"
 #include "syscall/internal.h"
 #include "syscall/mem.h"
