+++ src/syscall/net-absock.c
@@ -22,7 +22,7 @@
 
 #include "utils.h"
 
-#include "syscall/abi.h"
+#include "syscall/linux-wire.h"
 #include "syscall/internal.h"
 #include "syscall/net.h"
 #include "syscall/net-abi.h"
@@ -31,6 +31,7 @@
 
 #define ABSOCK_MAX_ENTRIES 64
 #define ABSOCK_MAX_NAME 107
+
 /* Derived-name budget shared by both namers: a literal prefix capped at
  * ABSOCK_PREFIX_MAX bytes for debuggability, then ABSOCK_DIGEST_HEX hex
  * characters of 64-bit FNV-1a that never truncate.
@@ -39,8 +40,9 @@
 #define ABSOCK_DIGEST_HEX 16
 /* Linux struct sockaddr_un carries at most 108 sun_path bytes. */
 #define LINUX_UNIX_PATH_MAX 108
-/* Width of the leading sun_family field, so the offset at which sun_path
- * begins and the length arithmetic around it cannot drift apart.
+
+/* Width of the leading sun_family field, so the offset at which sun_path begins
+ * and the length arithmetic around it cannot drift apart.
  */
 #define LINUX_SA_FAMILY_LEN 2
 
@@ -60,10 +62,10 @@ static _Atomic uint64_t absock_namespace_id;
 static _Atomic uint32_t absock_autobind_counter;
 
 /* Shortening links this process minted, recorded so exit unlinks exactly its
- * own: links are per-process property (the pid-scoped name guarantees no
- * other process shares one), so no sweep over the shared directory is needed
- * and none may run, because a sweep cannot tell a sibling's live link from a
- * leftover. Guarded by absock_lock.
+ * own: links are per-process property (the pid-scoped name guarantees no other
+ * process shares one), so no sweep over the shared directory is needed and none
+ * may run, because a sweep cannot tell a sibling's live link from a leftover.
+ * Guarded by absock_lock.
  */
 #define ABSOCK_MAX_LINKS 64
 static char absock_links[ABSOCK_MAX_LINKS][104];
@@ -71,10 +73,10 @@ static int absock_link_count;
 
 static void absock_cleanup(void);
 
-/* Spell the namespace directory for @namespace_id. Shared so a reader can
- * name the directory without depending on this process being the one that
- * created it, which is not the same thing: fork spawns a fresh process that
- * inherits the id and reads back sockaddrs pointing into the shared dir.
+/* Spell the namespace directory for @namespace_id. Shared so a reader can name
+ * the directory without depending on this process being the one that created
+ * it, which is not the same thing: fork spawns a fresh process that inherits
+ * the id and reads back sockaddrs pointing into the shared dir.
  */
 static int absock_dir_format(char *out, size_t out_sz, uint64_t namespace_id)
 {
@@ -90,10 +92,10 @@ static int absock_ensure_dir_locked(void)
         struct stat st;
 
         /* The directory is removed by whichever namespace participant exits
-         * last, and this process may outlive that: its cached flag then
-         * points at a directory that is gone, and every later mint would die
-         * with ENOENT. Recreate instead; create_private_dir tolerates EEXIST
-         * and re-validates ownership, and the atexit hook is already armed.
+         * last, and this process may outlive that: its cached flag then points
+         * at a directory that is gone, and every later mint would die with
+         * ENOENT. Recreate instead; create_private_dir tolerates EEXIST and
+         * re-validates ownership, and the atexit hook is already armed.
          */
         if (lstat(absock_dir, &st) == 0)
             return 0;
@@ -105,6 +107,7 @@ static int absock_ensure_dir_locked(void)
         atomic_store(&absock_namespace_id, namespace_id);
     }
     absock_dir_format(absock_dir, sizeof(absock_dir), namespace_id);
+
     /* The namespace-id path is guessable; create_private_dir rejects a
      * pre-planted symlink or foreign-owned directory in world-writable /tmp.
      */
@@ -156,8 +159,8 @@ void absock_encode_name(const char *dir,
             pos += (size_t) snprintf(out + pos, out_sz - pos, "%02x", name[i]);
     } else {
         /* The digest never truncates: the literal prefix shrinks to fit
-         * instead, because the prefix is debuggability while the digest is
-         * the only collision resistance a long name gets.
+         * instead, because the prefix is debuggability while the digest is the
+         * only collision resistance a long name gets.
          */
         size_t avail = out_sz > pos + ABSOCK_DIGEST_HEX + 1
                            ? out_sz - pos - ABSOCK_DIGEST_HEX - 1
@@ -295,8 +298,8 @@ static int absock_build_sun(const char *fs_path,
 
 /* Remember a minted link so absock_cleanup can unlink it. Full-table overflow
  * drops the record, never the link: an untracked link leaks until the last
- * participant's rmdir at worst, while refusing the mint would fail a live
- * bind or connect over bookkeeping.
+ * participant's rmdir at worst, while refusing the mint would fail a live bind
+ * or connect over bookkeeping.
  */
 static void absock_record_link_locked(const char *link_path)
 {
@@ -312,9 +315,9 @@ static void absock_record_link_locked(const char *link_path)
 
 /* Point a short symlink in the private absock dir at an over-long translated
  * socket path so it fits sun_path. bind(2) through a dangling symlink creates
- * the socket at the target and connect(2) follows it (probed on macOS 15).
- * The link name is pid-scoped (absock_link_name), so no two live processes
- * ever share one: a connect re-derives its own link, never a sibling's.
+ * the socket at the target and connect(2) follows it (probed on macOS 15). The
+ * link name is pid-scoped (absock_link_name), so no two live processes ever
+ * share one: a connect re-derives its own link, never a sibling's.
  */
 static int absock_shorten_path(const char *host_path, char *out, size_t out_sz)
 {
@@ -325,9 +328,10 @@ static int absock_shorten_path(const char *host_path, char *out, size_t out_sz)
         return -1;
     }
     int rc = symlink(host_path, out);
-    /* ENOENT here is the namespace directory vanishing between the ensure
-     * above and the symlink: another participant exited last and removed it.
-     * One re-ensure closes the race; a second loss just fails.
+
+    /* ENOENT here is the namespace directory vanishing between the ensure above
+     * and the symlink: another participant exited last and removed it. One
+     * re-ensure closes the race; a second loss just fails.
      */
     if (rc < 0 && errno == ENOENT && absock_ensure_dir_locked() == 0)
         rc = symlink(host_path, out);
@@ -343,9 +347,9 @@ static int absock_shorten_path(const char *host_path, char *out, size_t out_sz)
              * existing[0..n-1] on success.
              */
             memcmp(existing, host_path, (size_t) n)) {
-            /* A mismatched entry under a pid-scoped name is a leftover from
-             * a crashed run whose pid was recycled: no live process can own
-             * it, so replacing it is safe.
+            /* A mismatched entry under a pid-scoped name is a leftover from a
+             * crashed run whose pid was recycled: no live process can own it,
+             * so replacing it is safe.
              */
             (void) unlink(out);
             if (symlink(host_path, out) < 0) {
@@ -360,17 +364,17 @@ static int absock_shorten_path(const char *host_path, char *out, size_t out_sz)
 }
 
 /* Reverse-map one returned pathname AF_UNIX address to its guest spelling.
- * Returns the Linux sockaddr length when the address was rewritten, or -1
- * when it is not a translated pathname and the generic converter should run.
+ * Returns the Linux sockaddr length when the address was rewritten, or -1 when
+ * it is not a translated pathname and the generic converter should run.
  */
 static int absock_sockaddr_un_from_mac(const struct sockaddr_un *sun,
                                        uint32_t mac_len,
                                        uint8_t *linux_sa,
                                        uint32_t linux_sa_size)
 {
-    /* Bound every read by mac_len: macOS may fill all 104 sun_path bytes
-     * with no terminator, and only mac_len bytes of the caller's
-     * sockaddr_storage are initialized.
+    /* Bound every read by mac_len: macOS may fill all 104 sun_path bytes with
+     * no terminator, and only mac_len bytes of the caller's sockaddr_storage
+     * are initialized.
      */
     size_t sp_max = mac_len - offsetof(struct sockaddr_un, sun_path);
     if (sp_max > sizeof(sun->sun_path))
@@ -382,13 +386,14 @@ static int absock_sockaddr_un_from_mac(const struct sockaddr_un *sun,
     memcpy(mac_path, sun->sun_path, sp_len);
     mac_path[sp_len] = '\0';
 
-    /* Undo the over-length shortening symlink, then map the host path back
-     * to the guest namespace so the guest reads back the spelling it bound
-     * or connected with, not the sysroot-prefixed (and possibly escaped)
-     * host path.
+    /* Undo the over-length shortening symlink, then map the host path back to
+     * the guest namespace so the guest reads back the spelling it bound or
+     * connected with, not the sysroot-prefixed (and possibly escaped) host
+     * path.
      */
     char host_path[LINUX_PATH_MAX];
     str_copy_trunc(host_path, mac_path, sizeof(host_path));
+
     /* Compare against the whole directory component, not a byte prefix of it:
      * namespace 1234 would otherwise claim the paths of namespace 12345 and
      * readlink a directory it does not own.
@@ -412,10 +417,10 @@ static int absock_sockaddr_un_from_mac(const struct sockaddr_un *sun,
         !strcmp(guest_path, mac_path))
         return -1;
 
-    /* Write the Linux sockaddr directly: the guest may have bound a
-     * Linux-legal name longer than the 103 usable bytes of a macOS
-     * sun_path, and rebuilding a mac sockaddr first would fail for exactly
-     * the paths the shortening symlink serves.
+    /* Write the Linux sockaddr directly: the guest may have bound a Linux-legal
+     * name longer than the 103 usable bytes of a macOS sun_path, and rebuilding
+     * a mac sockaddr first would fail for exactly the paths the shortening
+     * symlink serves.
      */
     size_t glen = strlen(guest_path);
     if (glen > LINUX_UNIX_PATH_MAX || linux_sa_size < LINUX_SA_FAMILY_LEN)
@@ -451,9 +456,9 @@ int net_sockaddr_from_mac(const struct sockaddr *mac_sa,
 
 /* linux_errno() reports an int64_t syscall result, while these converters
  * return int to match linux_to_mac_sockaddr, the sibling they stand in for.
- * Every Linux errno is well under 4096 in magnitude, so the narrowing is
- * exact; it is spelled out rather than left implicit so it does not read as
- * an accident, and so -Wnarrowing-style analysis stays quiet about it.
+ * Every Linux errno is well under 4096 in magnitude, so the narrowing is exact;
+ * it is spelled out rather than left implicit so it does not read as an
+ * accident, and so -Wnarrowing-style analysis stays quiet about it.
  */
 static int absock_errno(void)
 {
@@ -476,9 +481,9 @@ int net_sockaddr_to_mac(const uint8_t *linux_sa,
          * bytes would name the unrelated host-literal file. Linux permits an
          * unterminated sun_path, so bound the copy by addrlen. A bind uses
          * create semantics, which spell an absent case-protected leaf at its
-         * escape, so the socket file lands where stat and connect will look;
-         * a name already bound resolves to the occupied spelling and the
-         * host bind reports EADDRINUSE, matching Linux.
+         * escape, so the socket file lands where stat and connect will look; a
+         * name already bound resolves to the occupied spelling and the host
+         * bind reports EADDRINUSE, matching Linux.
          */
         char guest_path[LINUX_UNIX_PATH_MAX + 1];
         uint32_t plen = addrlen - LINUX_SA_FAMILY_LEN;
@@ -498,9 +503,9 @@ int net_sockaddr_to_mac(const uint8_t *linux_sa,
          * take a sockaddr rather than a dirfd and at_flags, so it cannot ride
          * on an open flag here and is checked outright. Following a
          * guest-planted link would bind the socket at the link's target,
-         * outside the tree entirely, and would answer connect with ENOTSOCK
-         * for a host file that exists against ENOENT for one that does not,
-         * telling the guest whether any path exists. Both reach exactly what
+         * outside the tree entirely, and would answer connect with ENOTSOCK for
+         * a host file that exists against ENOENT for one that does not, telling
+         * the guest whether any path exists. Both reach exactly what
          * is_guest_system_path() denies the guest by name. See
          * dev_shm_resolve_path() in procemu.c.
          */
@@ -514,8 +519,8 @@ int net_sockaddr_to_mac(const uint8_t *linux_sa,
         const char *host_path = tx.host_path;
         if (strlen(host_path) >= sizeof(short_path)) {
             /* Surface the real failure (EACCES, EIO, ENOSPC, ...): the guest
-             * name is Linux-legal, so reporting ENAMETOOLONG would
-             * misattribute a symlink-layer error to the pathname length.
+             * name is Linux-legal, so reporting ENAMETOOLONG would misattribute
+             * a symlink-layer error to the pathname length.
              */
             if (absock_shorten_path(host_path, short_path, sizeof(short_path)) <
                 0)
@@ -621,8 +626,8 @@ void absock_bind_rollback(int idx)
 static void absock_cleanup(void)
 {
     /* Every process unlinks its own table-tracked sockets. A forked child
-     * starts from an empty table because fork hands over only the namespace
-     * id, so an entry here is never a sibling's.
+     * starts from an empty table because fork hands over only the namespace id,
+     * so an entry here is never a sibling's.
      */
     for (int i = 0; i < ABSOCK_MAX_ENTRIES; i++) {
         if (absock_table[i].active)
@@ -632,12 +637,12 @@ static void absock_cleanup(void)
     if (!absock_dir_created)
         return;
 
-    /* Shortening links are per-process property, recorded at mint time, so
-     * each process unlinks exactly its own whatever its relation to the
-     * namespace owner: a link another process still resolves through is
-     * never touched, and a child that minted links after the owner exited
-     * leaves nothing behind. An owner-only directory sweep gets both cases
-     * wrong, unlinking live siblings' links and skipping the orphan child's.
+    /* Shortening links are per-process property, recorded at mint time, so each
+     * process unlinks exactly its own whatever its relation to the namespace
+     * owner: a link another process still resolves through is never touched,
+     * and a child that minted links after the owner exited leaves nothing
+     * behind. An owner-only directory sweep gets both cases wrong, unlinking
+     * live siblings' links and skipping the orphan child's.
      */
     for (int i = 0; i < absock_link_count; i++)
         unlink(absock_links[i]);
